+++ src/main/java/ru/murlov/dto/CurrencyResponse.java
@@ -1,6 +1,6 @@
 package ru.murlov.dto;
 
-public record CurrencyResponse (int id,
+public record CurrencyResponse (Long id,
                                 String code,
                                 String name,
                                 String sign) {
