+++ src/main/java/ru/murlov/dto/ExchangeRateResponse.java
@@ -2,7 +2,7 @@ package ru.murlov.dto;
 
 import java.math.BigDecimal;
 
-public record ExchangeRateResponse (int id,
+public record ExchangeRateResponse (Long id,
                                     CurrencyResponse base_currency,
                                     CurrencyResponse target_currency,
                                     BigDecimal rate) {
